Refactor `bp/assignbppassport` package to `scanid/scannedqrcode`
