+++ app/src/main/java/org/simple/clinic/scanid/scannedqrcode/ScannedQrCodeEffectHandler.kt
@@ -1,4 +1,4 @@
-package org.simple.clinic.bp.assignbppassport
+package org.simple.clinic.scanid.scannedqrcode
 
 import com.spotify.mobius.rx2.RxMobius
 import dagger.assisted.Assisted
