+++ app/src/main/java/org/simple/clinic/scanid/scannedqrcode/ScannedQrCodeSheet.kt
@@ -1,4 +1,4 @@
-package org.simple.clinic.bp.assignbppassport
+package org.simple.clinic.scanid.scannedqrcode
 
 import android.content.Context
 import android.view.LayoutInflater
