+++ app/src/main/java/org/simple/clinic/scanid/scannedqrcode/ScannedQrCodeUi.kt
@@ -1,4 +1,4 @@
-package org.simple.clinic.bp.assignbppassport
+package org.simple.clinic.scanid.scannedqrcode
 
 interface ScannedQrCodeUi {
   fun showIndianNationalHealthIdValue()
