+++ app/src/main/java/org/simple/clinic/scanid/scannedqrcode/ScannedQrCodeUiActions.kt
@@ -1,4 +1,4 @@
-package org.simple.clinic.bp.assignbppassport
+package org.simple.clinic.scanid.scannedqrcode
 
 interface ScannedQrCodeUiActions {
   fun sendBpPassportResult(blankScannedQRCodeResult: BlankScannedQRCodeResult)
