+++ app/src/test/java/org/simple/clinic/scanid/scannedqrcode/ScannedQrCodeUiRendererTest.kt
@@ -1,12 +1,9 @@
-package org.simple.clinic.bp.assignpatientidentifier
+package org.simple.clinic.scanid.scannedqrcode
 
 import com.nhaarman.mockitokotlin2.mock
 import com.nhaarman.mockitokotlin2.verify
 import com.nhaarman.mockitokotlin2.verifyNoMoreInteractions
 import org.junit.Test
-import org.simple.clinic.bp.assignbppassport.ScannedQrCodeModel
-import org.simple.clinic.bp.assignbppassport.ScannedQrCodeUi
-import org.simple.clinic.bp.assignbppassport.ScannedQrCodeUiRenderer
 import org.simple.clinic.patient.businessid.Identifier
 import org.simple.clinic.patient.businessid.Identifier.IdentifierType.BpPassport
 import org.simple.clinic.patient.businessid.Identifier.IdentifierType.IndiaNationalHealthId
